+++ src/app/layout.tsx
@@ -4,6 +4,7 @@ import "./globals.css";
 import { cn } from "@/lib/utils";
 import RegisterSW from "@/components/RegisterSW";
 import InstallPWA from "@/components/InstallPWA";
+import { AuthProvider } from "@/contexts/AuthContext";
 
 const fontSans = FontSans({
   variable: "--font-sans",
@@ -34,7 +35,7 @@ export default function RootLayout({
   children: React.ReactNode;
 }>) {
   const BASE_PATH = process.env.NEXT_PUBLIC_BASE_PATH ?? "";
-  
+
   return (
     <html lang="en">
       <body
@@ -43,9 +44,9 @@ export default function RootLayout({
           fontSans.variable
         )}
       >
-        <main className="container mx-auto pt-5">
-        {children}
-        </main>
+        <AuthProvider>
+          {children}
+        </AuthProvider>
         <RegisterSW />
         <InstallPWA />
         <script